Multi-Walker/PPO_based/pyproject.toml
[project]
name = "5-stable"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "hydra-core>=1.3.2",
    "imageio>=2.36.0",
    "moviepy>=1.0.3",
    "pdm>=2.20.1",
    "pip==21",
    "protobuf==3.20.3",
    "setuptools==65.5.0",
    "tensorboard>=2.18.0",
    "wandb>=0.18.7",
    "wheel==0.38.0",
    "pettingzoo[sisl]>=1.24.3",
    "stable-baselines3[mpi]>=2.3.2",
    "supersuit>=3.9.3",
    "numpy>=2.1.3",
    "optuna>=4.0.0",
    "ipykernel>=6.29.5",
    "rich>=13.9.4",
    "hydra-joblib-launcher>=1.2.0",
    "hydra-ray-launcher>=1.2.1",
]
[tool.uv.sources]
torch = { index = "pytorch" }
benchmarl = { path = "BenchMARL" }

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl"
explicit = true